[dotfiles/.bash_logout,smlss.sh] Delete .bash_logout
I've kept this file for so long without really considering its actual
use. The one purpose of the file -- clearing the screen when the login
shell exits -- is so specific that it cannot justify the existance of a
whole file of its own in my home directory. Also, it seems it still does
that, even now that the file is deleted. I think I've kept this file all
the way from when I was on Debian, so it makes sense that its function
may not be necessary here on Arch. Goodbye, .bash_logout: good luck on
greener pastures.