.xinitrc: Don't query vpn info if NO_PROTONVPN_STATUSBAR is set
If I am unable to connect to ProtonVPN -- for example, if I am connected
to a network that blocks VPN connections -- then the status bar update
loop hangs for a long time (sometimes forever) waiting for protoninfo to
return. This variable will only be used in these circumstances.

TODO: add a signal or something so that I can change the value of
NO_PROTONVPN_STATUSBAR or somehow else tell .xinitrc that I do not want
to query for ProtonVPN status from within an already-existing dwm
session.