bash/aliases: Remove some editor aliases
If vi or vim are installed, I want to invoke them explicitly.

I use the alias v all of the time, so that stays.

When I use ed as an editor, I always like to have '-p:', but if I am
using it in as part of a pipe (for example,
`echo -e '2d\nwq' | ed TODO`), the prompt displays along with the
desired output.