.xinitrc: Toggle VPN status info by checking modify date
Previously, VPN status info was toggled by seeing if a certain file
existed, and then removing it. This worked, but it was difficult to
remember the path to the file because it never existed when you wanted
to toggle status info. This problem is fixed by instead touching a
constant file.