src/.xinitrc: Simplify batinfo
First, usage of the || operator is replaced with an if statement.
grep is called with -v instead of using an else or negating the call
with a !.

Complex control flow combining awk, sed, tr, non-standard echo,
and a case statement (using another call to acpi and awk) is replaced
with a single, commented inline awk script.